+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Nightly](https://github.com/raymond-u/lungo/compare/v0.5.0...HEAD)
 
+### Fixed
+
+- Generate Kratos secrets correctly
+
 ## [0.5.0](https://github.com/raymond-u/lungo/compare/v0.4.2...v0.5.0) - 2024-6-22
 
 ### Breaking Changes